+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.20.1 Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
-## [Unreleased]
+## [1.2.10]
 ### Added
  - Supports Giga Pattern Provider (Expanded AE).